+++ src/protocols/gg_2020/state_machine/keygen/messages/address.rs
@@ -7,6 +7,6 @@ use serde::{
 
 #[derive(Debug, Clone, Deserialize, Serialize, PartialEq)]
 pub enum Address {
-    Peer(u16),
+    Peer(usize),
     Broadcast,
 }
